Move koov support code into own directory under src/extensions.
